+++ index.js
@@ -2,5 +2,5 @@
 'use strict';
 
 module.exports = {
-  name: 'sl-ember-test-helpers'
+    name: 'sl-ember-test-helpers'
 };