+++ test-support/helpers/sl/synchronous.js
@@ -1,9 +1,11 @@
 import AjaxHelper from './synchronous/ajax';
 import contains from './synchronous/contains';
 import requires from './synchronous/requires';
+import * as globalLibraries from './synchronous/global-libraries';
 
 export {
     AjaxHelper,
     contains,
-    requires
+    requires,
+    globalLibraries
 };